backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
aiofiles==24.1.0
pillow==11.1.0
python-slugify==8.0.4
python-frontmatter==1.1.0
email-validator==2.2.0
bcrypt==4.0.1
